rename bookAction.js to books.js
