+++ src/redux/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
-import bookReducer from './books/bookAction';
+import bookReducer from './books/books';
 
 export const store = configureStore({
   reducer: {
